requirements.txt
clyngor==0.3.21
watchdog==0.9.0
lark-parser==0.7.5
terminaltables==3.1.0
randomcolor==0.4.4.5
ansicolors==1.1.8
pandas==0.25.3
argh==0.26.2